Minor tweak to summary
Daniel - just altered the summary slightly. See if you agree. I’m good
with the report - I think it’s in good shape for submission. If/when
you are comfortable - can you submit on our behalf?
(Best of luck with the prez tomorrow, I hope to catch the video
recording when I’m back at the end of the week)!

Fab working with you - thanks for all your hard work!
Jay